+++ src/reports/index.html
@@ -35,7 +35,7 @@
 <a href="#"><span class="badge badge-primary">Validate E2E function </span></a>
 </li>
 <li class="m-r-10">
-<a href="#"><span class="badge badge-primary">Dec 11, 2023 04:23:00 PM</span></a>
+<a href="#"><span class="badge badge-primary">Dec 12, 2023 11:25:54 AM</span></a>
 </li>
 </ul>
 </div>
@@ -86,7 +86,7 @@
           <div class="test-detail">
             <p class="name">testBlankPage</p>
             <p class="text-sm">
-              <span>4:23:08 PM</span> / <span>00:00:00:805</span>
+              <span>11:26:05 AM</span> / <span>00:00:00:618</span>
               <span class="badge fail-bg log float-right">Fail</span>
             </p>
           </div>
@@ -95,9 +95,9 @@
 <div class="p-v-10">
 <div class="info">
 <h5 class="test-status text-fail">testBlankPage</h5>
-<span class='badge badge-success'>12.11.2023 4:23:08 PM</span>
-<span class='badge badge-danger'>12.11.2023 4:23:09 PM</span>
-<span class='badge badge-default'>00:00:00:805</span>
+<span class='badge badge-success'>12.12.2023 11:26:05 AM</span>
+<span class='badge badge-danger'>12.12.2023 11:26:06 AM</span>
+<span class='badge badge-default'>00:00:00:618</span>
 &middot; <span class='uri-anchor badge badge-default'>#test-id=1</span>
 <span title='Skip to the next failed step' class='badge badge-danger pointer float-right ne ml-1'><i class="fa fa-fast-forward"></i></span>
 <span title='Collapse all nodes' class='badge badge-default pointer float-right ct ml-1'><i class="fa fa-compress"></i></span>
@@ -110,25 +110,25 @@
   <tbody>
       <tr class="event-row">
         <td><span class="badge log pass-bg">Pass</span></td>
-        <td>4:23:09 PM</td>
+        <td>11:26:06 AM</td>
         <td>
           testBlankPage pass
         </td>
       </tr>
       <tr class="event-row">
         <td><span class="badge log fail-bg">Fail</span></td>
-        <td>4:23:09 PM</td>
+        <td>11:26:06 AM</td>
         <td>
           testBlankPage fail
         </td>
       </tr>
       <tr class="event-row">
         <td><span class="badge log fail-bg">Fail</span></td>
-        <td>4:23:09 PM</td>
+        <td>11:26:06 AM</td>
         <td>
           <textarea readonly class="code-block">java.lang.NullPointerException
-	at wikipedia.resources.Utils.takeScreenShot(Utils.java:35)
-	at wikipedia.Listener.onTestSuccess(Listener.java:42)
+	at utils.Utils.takeScreenShot(Utils.java:35)
+	at extentReportListener.TestNGListener.onTestSuccess(TestNGListener.java:44)
 	at org.testng.internal.TestListenerHelper.runTestListeners(TestListenerHelper.java:112)
 	at org.testng.internal.invokers.TestInvoker.runTestResultListener(TestInvoker.java:262)
 	at org.testng.internal.invokers.TestInvoker.invokeMethod(TestInvoker.java:751)
@@ -203,7 +203,7 @@
 <tbody>
 <tr class="tag-test-status" status="fail">
 <td><span class="badge log fail-bg">Fail</span></td>
-<td>16:23:08 PM</td>
+<td>11:26:05 AM</td>
 <td>
 <a href="#" class="linked" test-id='1' id='1'>testBlankPage</a>
 </td>
@@ -225,13 +225,13 @@
 <div class="col-md-3">
 <div class="card"><div class="card-body">
 <p class="m-b-0">Started</p>
-<h3>Dec 11, 2023 04:23:00 PM</h3>
+<h3>Dec 12, 2023 11:25:54 AM</h3>
 </div></div>
 </div>
 <div class="col-md-3">
 <div class="card"><div class="card-body">
 <p class="m-b-0">Ended</p>
-<h3>Dec 11, 2023 04:23:09 PM</h3>
+<h3>Dec 12, 2023 11:26:06 AM</h3>
 </div></div>
 </div>
 <div class="col-md-3">
@@ -301,7 +301,7 @@
 </div></div>
 <script>
 var timeline = {
-"testBlankPage":0.805
+"testBlankPage":0.618
 };
 </script>
 <div class="row">
